Code Review Project 4
Good job on your project! All the functions are clear and easy to find, and I like that you neatly separate file in different folders so it's clearly organized. The search functionality and UI looked great as well. One small advice that on the readme, maybe add instructions to run front end dev server.